requirements.txt
# StatementXL Version 2 - Dependencies
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.25
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Validation & Settings
pydantic>=2.6.0
pydantic-settings>=2.1.0

# PDF Processing
pdfplumber>=0.10.0
pytesseract>=0.3.10
camelot-py[cv]>=0.11.0
opencv-python-headless>=4.9.0
ghostscript>=0.7
PyMuPDF>=1.23.0
Pillow>=10.0.0

# Logging
structlog>=24.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.0
